requirements.txt
fastapi
uvicorn
streamlit
pydantic
python-dotenv
google-genai
